// packages/net/src/jsonrpc/_types/SuccessResponse.d.ts
/**
* Source file:  "SuccessResponse.json"
* Location:     "/home/alykoshin/projects/dev/082-mt-node-red/experiments/ws-client-server/lib/jsonrpc"
* Generated by: "_utils/build.sh"
* Date:         "2020-09-01T01:57:38+03:00" 
*/

/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SuccessResponse {
  /**
   * A String specifying the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: string;
  /**
   * This member is REQUIRED on success.
   * This member MUST NOT exist if there was an error invoking the method.
   * The value of this member is determined by the method invoked on the Server.
   */
  result: {
    [k: string]: unknown;
  };
  /**
   * This member is REQUIRED.
   * It MUST be the same as the value of the id member in the Request Object.
   * If there was an error in detecting the id in the Request object (e.g. Parse error/Invalid Request), it MUST be Null.
   */
  id: string | number | null;
}